+++ src/features/amUI/common/AccessPackageList/AreaList.module.css
